# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for property listings with geospatial search.",
        "title": "Property Marketplace API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/agents": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "List all agents",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.DataListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Create an agent",
                "parameters": [
                    {
                        "description": "Agent payload",
                        "name": "agent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.CreateAgentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.DataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/agents/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Get an agent by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.DataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Update an agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated agent payload",
                        "name": "agent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.UpdateAgentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.DataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Delete an agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_agent_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_agent_handler.AgentResponse": {
            "type": "object",
            "properties": {
                "agency": {
                    "type": "string",
                    "example": "Realty NG"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "ada@realty.ng"
                },
                "id": {
                    "type": "string",
                    "example": "a1000000-0000-0000-0000-000000000001"
                },
                "name": {
                    "type": "string",
                    "example": "Ada Okafor"
                },
                "phone": {
                    "type": "string",
                    "example": "+2348011111111"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_agent_handler.CreateAgentRequest": {
            "type": "object",
            "required": [
                "agency",
                "email",
                "name",
                "phone"
            ],
            "properties": {
                "agency": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Realty NG"
                },
                "email": {
                    "type": "string",
                    "example": "ada@realty.ng"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Ada Okafor"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 7,
                    "example": "+2348011111111"
                }
            }
        },
        "internal_agent_handler.DataListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_agent_handler.AgentResponse"
                    }
                },
                "pagination": {}
            }
        },
        "internal_agent_handler.DataResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_agent_handler.AgentResponse"
                }
            }
        },
        "internal_agent_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "agent not found"
                }
            }
        },
        "internal_agent_handler.UpdateAgentRequest": {
            "type": "object",
            "required": [
                "agency",
                "email",
                "name",
                "phone"
            ],
            "properties": {
                "agency": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Realty NG"
                },
                "email": {
                    "type": "string",
                    "example": "ada@realty.ng"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Ada Okafor"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 7,
                    "example": "+2348011111111"
                }
            }
        },
        "internal_agent_handler.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_agent_handler.ValidationFieldError"
                    }
                }
            }
        },
        "internal_agent_handler.ValidationFieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "email"
                },
                "message": {
                    "type": "string",
                    "example": "must be a valid email address"
                }
            }
        },
...
    },
...
}